verb-forms: Add lesson filter to `RunSpec`.
Now lesson filter may be specified in yaml in `filter` field. It may be a
single lesson number

    filter: 10

or a range

    filter: { from: 10, till: 15 }

One of `from` and `till` may be omitted making an incomplete range (the missed
range end will be the first or the last available lesson respectively).

Note, that for now, it's impossible to include verbs _not_ assigned to any
lesson.